GLOSSARY-53 Add ability to sync multilang Glossary with deepl glossary
* Move translator service glossary management to translator project. This is needed as we can't depends on glossary application as it will create loop dependency
* Rename componenet for more consistency
