+++ application-glossary-machine-translation/application-glossary-machine-translation-api/src/main/resources/META-INF/components.txt
@@ -1,0 +1,1 @@
+org.xwiki.contrib.glossary.machineTranslation.TranslationGlossaryScriptService
