+++ application-glossary-machine-translation/application-glossary-machine-translation-ui/src/main/resources/Glossary/Code/MachineTranslation/MachineTranslationTranslations.xml
@@ -1,0 +1,89 @@
+<?xml version="1.1" encoding="UTF-8"?>
+
+<!--
+ * See the NOTICE file distributed with this work for additional
+ * information regarding copyright ownership.
+ *
+ * This is free software; you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation; either version 2.1 of
+ * the License, or (at your option) any later version.
+ *
+ * This software is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this software; if not, write to the Free
+ * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
+ * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
+-->
+
+<xwikidoc version="1.4" reference="Glossary.Code.MachineTranslation.MachineTranslationTranslations" locale="">
+  <web>Glossary.Code.MachineTranslation</web>
+  <name>MachineTranslationTranslations</name>
+  <language/>
+  <defaultLanguage>en</defaultLanguage>
+  <translation>0</translation>
+  <creator>xwiki:XWiki.Admin</creator>
+  <parent>Glossary.Code.MachineTranslation.WebHome</parent>
+  <author>xwiki:XWiki.Admin</author>
+  <contentAuthor>xwiki:XWiki.Admin</contentAuthor>
+  <version>1.1</version>
+  <title>MachineTranslationTranslations</title>
+  <comment/>
+  <minorEdit>false</minorEdit>
+  <syntaxId>plain/1.0</syntaxId>
+  <hidden>true</hidden>
+  <content>glossary.machineTranslation.explorer.title=Translation Glossaries Explorer
+glossary.machineTranslation.explorer.forceSync=Force synchronization
+glossary.machineTranslation.explorer.glossary=Glossary
+glossary.machineTranslation.explorer.glossaries=Glossaries
+glossary.machineTranslation.explorer.glossaryId=ID
+glossary.machineTranslation.explorer.glossaryName=Name
+glossary.machineTranslation.explorer.sourceLanguage=Source language
+glossary.machineTranslation.explorer.targetLanguage=Target language
+glossary.machineTranslation.explorer.isReady=Is ready ?
+glossary.machineTranslation.explorer.entryCount=Entry count
+glossary.machineTranslation.explorer.supportedLanguagePairs=Supported language pairs
+glossary.machineTranslation.usage.charCount=Character count :
+glossary.machineTranslation.usage.charLimit=Character limit :</content>
+  <object>
+    <name>Glossary.Code.MachineTranslation.MachineTranslationTranslations</name>
+    <number>0</number>
+    <className>XWiki.TranslationDocumentClass</className>
+    <guid>ccd1b16f-0b93-4957-b24d-c92d783e588d</guid>
+    <class>
+      <name>XWiki.TranslationDocumentClass</name>
+      <customClass/>
+      <customMapping/>
+      <defaultViewSheet/>
+      <defaultEditSheet/>
+      <defaultWeb/>
+      <nameField/>
+      <validationScript/>
+      <scope>
+        <cache>0</cache>
+        <disabled>0</disabled>
+        <displayType>select</displayType>
+        <freeText>forbidden</freeText>
+        <largeStorage>0</largeStorage>
+        <multiSelect>0</multiSelect>
+        <name>scope</name>
+        <number>1</number>
+        <prettyName>Scope</prettyName>
+        <relationalStorage>0</relationalStorage>
+        <separator> </separator>
+        <separators>|, </separators>
+        <size>1</size>
+        <unmodifiable>0</unmodifiable>
+        <values>GLOBAL|WIKI|USER|ON_DEMAND</values>
+        <classType>com.xpn.xwiki.objects.classes.StaticListClass</classType>
+      </scope>
+    </class>
+    <property>
+      <scope>WIKI</scope>
+    </property>
+  </object>
+</xwikidoc>
