+++ application-glossary-machine-translation/application-glossary-machine-translation-ui/src/main/resources/Glossary/Code/MachineTranslation/WebHome.xml
@@ -1,4 +1,6 @@
-/*
+<?xml version="1.1" encoding="UTF-8"?>
+
+<!--
  * See the NOTICE file distributed with this work for additional
  * information regarding copyright ownership.
  *
@@ -16,33 +18,25 @@
  * License along with this software; if not, write to the Free
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
- */
-package org.xwiki.contrib.glossary.script;
+-->
 
-import javax.inject.Inject;
-import javax.inject.Named;
-import javax.inject.Singleton;
+<xwikidoc version="1.4" reference="Glossary.Code.MachineTranslation.WebHome" locale="">
+  <web>Glossary.Code.MachineTranslation</web>
+  <name>WebHome</name>
+  <language/>
+  <defaultLanguage/>
+  <translation>0</translation>
+  <creator>xwiki:XWiki.Admin</creator>
+  <parent>Glossary.Code.WebHome</parent>
+  <author>xwiki:XWiki.Admin</author>
+  <contentAuthor>xwiki:XWiki.Admin</contentAuthor>
+  <version>1.1</version>
+  <title>Machine translation glossary</title>
+  <comment/>
+  <minorEdit>false</minorEdit>
+  <syntaxId>xwiki/2.1</syntaxId>
+  <hidden>true</hidden>
+  <content>Code specific to the glossary integration with the translator application.
 
-import org.xwiki.component.annotation.Component;
-import org.xwiki.script.service.ScriptService;
-import org.xwiki.script.service.ScriptServiceManager;
-
-/**
- * Provides Component-specific Scripting APIs.
- * Used to provide the access to the service 'services.glossary.translation
- */
-@Component
-@Named(GlossaryScriptService.HINT)
-@Singleton
-public class GlossaryScriptService implements ScriptService
-{
-    final static String HINT = "glossary";
-
-    @Inject
-    private ScriptServiceManager scriptServiceManager;
-
-    public <S extends ScriptService> S get(String serviceName)
-    {
-        return (S) this.scriptServiceManager.get(HINT + "." + serviceName);
-    }
-}
+{{children/}}</content>
+</xwikidoc>
