--- application-glossary-translator/application-glossary-translator-api/src/main/resources/META-INF/components.txt
@@ -1,1 +1,0 @@
-org.xwiki.contrib.glossary.translator.TranslationGlossaryScriptService
