--- application-glossary-translator/application-glossary-translator-ui/src/main/resources/Glossary/Code/Translator/GlossaryTranslatorTranslations.xml
@@ -1,96 +1,0 @@
-<?xml version="1.1" encoding="UTF-8"?>
-
-<!--
- * See the NOTICE file distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
--->
-
-<xwikidoc version="1.4" reference="Glossary.Code.Translator.GlossaryTranslatorTranslations" locale="">
-  <web>Glossary.Code.Translator</web>
-  <name>GlossaryTranslatorTranslations</name>
-  <language/>
-  <defaultLanguage>en</defaultLanguage>
-  <translation>0</translation>
-  <creator>xwiki:XWiki.Admin</creator>
-  <parent>WebHome</parent>
-  <author>xwiki:XWiki.Admin</author>
-  <contentAuthor>xwiki:XWiki.Admin</contentAuthor>
-  <version>1.1</version>
-  <title>GlossaryTranslatorTranslations</title>
-  <comment/>
-  <minorEdit>false</minorEdit>
-  <syntaxId>plain/1.0</syntaxId>
-  <hidden>true</hidden>
-  <content>
-
-admin.glossarytranslator=Translation Glossary
-
-glossary.translator.explorer.title=Translation Glossaries Explorer
-glossary.translator.explorer.forceSync=Force synchronization
-glossary.translator.explorer.glossary=Glossary
-glossary.translator.explorer.glossaries=Glossaries
-glossary.translator.explorer.glossaryId=ID
-glossary.translator.explorer.glossaryName=Name
-glossary.translator.explorer.sourceLanguage=Source language
-glossary.translator.explorer.targetLanguage=Target language
-glossary.translator.explorer.isReady=Is ready ?
-glossary.translator.explorer.entryCount=Entry count
-glossary.translator.explorer.supportedLanguagePairs=Supported language pairs
-glossary.translator.usage.charCount=Character count :
-glossary.translator.usage.charLimit=Character limit :
-
-Glossary.Code.Translator.TranslatorGlossaryConfigurationClass_translatorGlossaryNamePrefix=Glossary name prefix
-Glossary.Code.Translator.TranslatorGlossaryConfigurationClass_translatorGlossaryNamePrefix.hint=Prefix of glossay name into the translator provider.</content>
-  <object>
-    <name>Glossary.Code.Translator.GlossaryTranslatorTranslations</name>
-    <number>0</number>
-    <className>XWiki.TranslationDocumentClass</className>
-    <guid>ccd1b16f-0b93-4957-b24d-c92d783e588d</guid>
-    <class>
-      <name>XWiki.TranslationDocumentClass</name>
-      <customClass/>
-      <customMapping/>
-      <defaultViewSheet/>
-      <defaultEditSheet/>
-      <defaultWeb/>
-      <nameField/>
-      <validationScript/>
-      <scope>
-        <cache>0</cache>
-        <disabled>0</disabled>
-        <displayType>select</displayType>
-        <freeText>forbidden</freeText>
-        <largeStorage>0</largeStorage>
-        <multiSelect>0</multiSelect>
-        <name>scope</name>
-        <number>1</number>
-        <prettyName>Scope</prettyName>
-        <relationalStorage>0</relationalStorage>
-        <separator> </separator>
-        <separators>|, </separators>
-        <size>1</size>
-        <unmodifiable>0</unmodifiable>
-        <values>GLOBAL|WIKI|USER|ON_DEMAND</values>
-        <classType>com.xpn.xwiki.objects.classes.StaticListClass</classType>
-      </scope>
-    </class>
-    <property>
-      <scope>WIKI</scope>
-    </property>
-  </object>
-</xwikidoc>
