--- application-glossary-translator/application-glossary-translator-ui/src/main/resources/Glossary/Code/Translator/SynchronizeGlossaries.xml
@@ -1,168 +1,0 @@
-<?xml version="1.1" encoding="UTF-8"?>
-
-<!--
- * See the NOTICE file distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
--->
-
-<xwikidoc version="1.4" reference="Glossary.Code.Translator.SynchronizeGlossaries" locale="">
-  <web>Glossary.Code.Translator</web>
-  <name>SynchronizeGlossaries</name>
-  <language/>
-  <defaultLanguage/>
-  <translation>0</translation>
-  <creator>xwiki:XWiki.Admin</creator>
-  <parent>WebHome</parent>
-  <author>xwiki:XWiki.Admin</author>
-  <contentAuthor>xwiki:XWiki.Admin</contentAuthor>
-  <version>1.1</version>
-  <title>Synchronize Glossaries</title>
-  <comment/>
-  <minorEdit>false</minorEdit>
-  <syntaxId>xwiki/2.1</syntaxId>
-  <hidden>true</hidden>
-  <content/>
-  <object>
-    <name>Glossary.Code.Translator.SynchronizeGlossaries</name>
-    <number>0</number>
-    <className>XWiki.SchedulerJobClass</className>
-    <guid>290f402c-af51-478e-abbd-d9d8a5994ab0</guid>
-    <class>
-      <name>XWiki.SchedulerJobClass</name>
-      <customClass/>
-      <customMapping/>
-      <defaultViewSheet/>
-      <defaultEditSheet/>
-      <defaultWeb/>
-      <nameField/>
-      <validationScript/>
-      <contextDatabase>
-        <disabled>0</disabled>
-        <name>contextDatabase</name>
-        <number>9</number>
-        <prettyName>Job execution context database</prettyName>
-        <size>30</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </contextDatabase>
-      <contextLang>
-        <disabled>0</disabled>
-        <name>contextLang</name>
-        <number>8</number>
-        <prettyName>Job execution context lang</prettyName>
-        <size>30</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </contextLang>
-      <contextUser>
-        <disabled>0</disabled>
-        <name>contextUser</name>
-        <number>7</number>
-        <prettyName>Job execution context user</prettyName>
-        <size>30</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </contextUser>
-      <cron>
-        <disabled>0</disabled>
-        <name>cron</name>
-        <number>5</number>
-        <prettyName>Cron Expression</prettyName>
-        <size>30</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </cron>
-      <jobClass>
-        <disabled>0</disabled>
-        <name>jobClass</name>
-        <number>3</number>
-        <prettyName>Job Class</prettyName>
-        <size>60</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </jobClass>
-      <jobDescription>
-        <disabled>0</disabled>
-        <name>jobDescription</name>
-        <number>2</number>
-        <prettyName>Job Description</prettyName>
-        <rows>10</rows>
-        <size>45</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.TextAreaClass</classType>
-      </jobDescription>
-      <jobName>
-        <disabled>0</disabled>
-        <name>jobName</name>
-        <number>1</number>
-        <prettyName>Job Name</prettyName>
-        <size>60</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </jobName>
-      <script>
-        <contenttype>PureText</contenttype>
-        <disabled>0</disabled>
-        <editor>PureText</editor>
-        <name>script</name>
-        <number>6</number>
-        <prettyName>Job Script</prettyName>
-        <rows>10</rows>
-        <size>60</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.TextAreaClass</classType>
-      </script>
-      <status>
-        <disabled>0</disabled>
-        <name>status</name>
-        <number>4</number>
-        <prettyName>Status</prettyName>
-        <size>30</size>
-        <unmodifiable>0</unmodifiable>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </status>
-    </class>
-    <property>
-      <contextDatabase>xwiki</contextDatabase>
-    </property>
-    <property>
-      <contextLang>en</contextLang>
-    </property>
-    <property>
-      <contextUser>superadmin</contextUser>
-    </property>
-    <property>
-      <cron>0 15 1 * * ? </cron>
-    </property>
-    <property>
-      <jobClass>com.xpn.xwiki.plugin.scheduler.GroovyJob</jobClass>
-    </property>
-    <property>
-      <jobDescription>Will synchronize the contents of the XWiki Glossaries with the translator service</jobDescription>
-    </property>
-    <property>
-      <jobName>Synchronize Glossaries</jobName>
-    </property>
-    <property>
-      <script>services.glossary.translator.synchronizeGlossaries()</script>
-    </property>
-    <property>
-      <status>Normal</status>
-    </property>
-  </object>
-</xwikidoc>
