--- application-glossary-translator/application-glossary-translator-ui/src/main/resources/Glossary/TranslatorGlossaryConfiguration.xml
@@ -1,74 +1,0 @@
-<?xml version="1.1" encoding="UTF-8"?>
-
-<!--
- * See the NOTICE file distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This is free software; you can redistribute it and/or modify it
- * under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation; either version 2.1 of
- * the License, or (at your option) any later version.
- *
- * This software is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public
- * License along with this software; if not, write to the Free
- * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
- * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
--->
-
-<xwikidoc version="1.5" reference="Glossary.TranslatorGlossaryConfiguration" locale="">
-  <web>Glossary</web>
-  <name>TranslatorGlossaryConfiguration</name>
-  <language/>
-  <defaultLanguage/>
-  <translation>0</translation>
-  <creator>xwiki:XWiki.Admin</creator>
-  <parent>WebHome</parent>
-  <author>xwiki:XWiki.Admin</author>
-  <originalMetadataAuthor>XWiki.Admin</originalMetadataAuthor>
-  <contentAuthor>xwiki:XWiki.Admin</contentAuthor>
-  <version>1.1</version>
-  <title/>
-  <comment/>
-  <minorEdit>false</minorEdit>
-  <syntaxId>xwiki/2.1</syntaxId>
-  <hidden>true</hidden>
-  <content/>
-  <object>
-    <name>Glossary.TranslatorGlossaryConfiguration</name>
-    <number>0</number>
-    <className>Glossary.Code.Translator.TranslatorGlossaryConfigurationClass</className>
-    <guid>821927ff-359a-4e80-ac64-9e8c5a8eaa05</guid>
-    <class>
-      <name>Glossary.Code.Translator.TranslatorGlossaryConfigurationClass</name>
-      <customClass/>
-      <customMapping/>
-      <defaultViewSheet/>
-      <defaultEditSheet/>
-      <defaultWeb/>
-      <nameField/>
-      <validationScript/>
-      <translatorGlossaryNamePrefix>
-        <customDisplay/>
-        <disabled>0</disabled>
-        <hint/>
-        <name>translatorGlossaryNamePrefix</name>
-        <number>1</number>
-        <picker>1</picker>
-        <prettyName>translatorGlossaryNamePrefix</prettyName>
-        <size>30</size>
-        <unmodifiable>0</unmodifiable>
-        <validationMessage/>
-        <validationRegExp/>
-        <classType>com.xpn.xwiki.objects.classes.StringClass</classType>
-      </translatorGlossaryNamePrefix>
-    </class>
-    <property>
-      <translatorGlossaryNamePrefix/>
-    </property>
-  </object>
-</xwikidoc>
